+++ appender/PopUpAppender.ts
@@ -18,18 +18,16 @@ class PopUpAppender extends ConsoleAppender {
         reopenWhenClosed: false,
         maxMessages: <number><any>null,
         showCommandLine: true,
-        commandLineObjectExpansionDepth: 1,
         showHideButton: false,
         showCloseButton: true,
-        useDocumentWrite: true
     };
     public name = "PopUpAppender";
 
 
-    constructor(lazyInit?: boolean, initiallyMinimized?: boolean, useDocumentWrite?: boolean, width?: string | number, height?: string | number, opts?: Log4Ts.AppenderOptions) {
+    constructor(lazyInit?: boolean, initiallyMinimized?: boolean, width?: string | number, height?: string | number, opts?: Log4Ts.AppenderOptions) {
         super(opts);
 
-        this.create(false, null, lazyInit, initiallyMinimized, useDocumentWrite, width, height, this.defaults.focusPopUp);
+        this.create(false, null, lazyInit, initiallyMinimized, width, height, this.defaults.focusPopUp);
     }